Adding driver.cpp
Adding initial driver source file as specified in Project Part 1.  This version will read in strings from stdin, ignoring whitespace.  Any strings will be looked-up in the symbol table and print the address of the symbol in the table.  The only deviation from the part 1 specification is that the program doesn't terminate after the initial read:  the user will need to manually terminate in this version.