Old File Removal and Exit Code Update
The group has decided to remove our archives to clean up the repository.

This commit also updates the exit code to use a non-zero number since
exit with a non-zero integer parameter represents an errored out exit.
That is what that line is supposed to do in the first place, since out
of memory means a fatal error.