File I/O Update
Compiler program will now take in a file name directly from the command line.  The file name is assumed to be the last argument passed to the compiler (so DEBUG flags come first).  The compiler will output to a .q file that has the same name as the passed in file after evaluating the AST.